git-svn-id: http://ownageclan.com/svn/war3source@274 e104397d-def3-435e-8314-01917568762b

+++ extensionsource/war3ext/updatefile.cpp
@@ -5,12 +5,12 @@
 void updatefile(string file){
 	for(int breakme=0;breakme<1;breakme++){
 	int pos=file.find('\n');
-	if(pos!=string::npos){
+	if(pos!=(int)string::npos){
 		file.erase ( pos, 1);
 		//cout<<"FOUND slashN"<<endl;
 	}
 	pos=file.find('\r');
-	if(pos!=string::npos){
+	if(pos!=(int)string::npos){
 		file.erase ( pos, 1);
 		//cout<<"FOUND slashR"<<endl;
 	}
